test case updates
still not sure what is causing the issue. it seems like some of the threads are not getting a properly updated value when it is changed in another thread but the mutex should be ensuring exclusive access. I made a separate module to test in because I thought the cargo test could have been doing something weird but still seeing similar behavior. the custom test creates a new file that will display rough timing information to hopefully show when the id is generated relative to the test start