FOUND THE BUG
as I was writing out some comments on the process I realized that the timestamp and check are also dependent on when the mutex is acquired. moved it into the critical section and the bugs stopped.

added in more documentation for things.
added in some convenience functions for Snowflake and Snowcloud
moved tests from custom_test into a file in src since the testing framework was not causing issues

going to look into using SystemTime and Duration for calculations and see if it is more ergonomic and maybe faster?