updated to use SystemTime and Duration
per a suggestion from a friend, moved to using SystemTime for storing the desired EPOCH from UNIX EPOCH and using Duration for getting snowflake timestamp and storing the previous timestamp. got some performance improvements when generating flakes since it is no longer doing as much conversions on each call. get better accuracy of when the flake is created but the accuracy is lost when converting to i64's, creating a flake from an i64, or creating a flake from parts.