generic traits and blocking_next_id
added in two new traits IdGenerator and NextAvailId. with these it should be possible to create other functions to help with generating ids outside of the base class. currently being used with blocking_next_id formerly spin_next_id.
also found a calcuation error when returning the time to the next millisecond. previousely was just returning the nanoseconds and not actually getting the time to the next one. it popped up when testing blocking_next_id since it should not take more than 2 attempts to get an id in a single thread.