separated modules
separated out the modules to allow for the ability to pull only what you
need and to better allow for additions outside of snowcloud.

updated the snowflake build process to use a builder trait. it will
allow for more custom behavior and allow for the generator to know when
to respond with an error. currently does not support having the builder
return an error but maybe in the future.

docs still need to be updated.